+++ ecs/storage/observer_index.h
@@ -19,14 +19,14 @@ typedef struct {
 void ecs_observer_index_init(ecs_observer_index_t *index);
 void ecs_observer_index_fini(ecs_observer_index_t *index);
 
-uint32_t ecs_observer_index_create(ecs_observer_index_t *index, const ecs_observer_desc_t *desc);
+uint16_t ecs_observer_index_create(ecs_observer_index_t *index, const ecs_observer_desc_t *desc);
 
 // Cache a freshly created observer onto every existing table it matches.
 void ecs_observer_index_match_tables(
     ecs_observer_index_t *index,
     ecs_table_t *tables,
     uint16_t table_count,
-    uint32_t observer_id
+    uint16_t observer_id
 );
 
 // Cache every existing observer that matches a freshly created table.
